requirements.txt
Django>=5.2
django-allauth>=0.61.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
mysqlclient>=2.2.1
Pillow>=10.2.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
